+++ trixnity-messenger/src/webMain/kotlin/de/connect2x/trixnity/messenger/media/WebMediaPlayer.kt
@@ -38,29 +38,31 @@ import web.html.Preload
 import web.html.metadata
 import web.mediasession.MediaSessionAction
 import web.mediasession.MediaSessionActionDetails
+import web.mediasession.nexttrack
 import web.mediasession.pause
 import web.mediasession.play
+import web.mediasession.previoustrack
 import web.mediasession.seekbackward
 import web.mediasession.seekforward
 import web.mediasession.seekto
+import web.mediasession.skipad
+import web.mediasession.stop
 import web.navigator.navigator
 import web.url.URL
 
-class WebMediaPlayer(private val coroutineScope: CoroutineScope) : MediaPlayer {
+class WebMediaPlayer(private val audioContext: AudioContext, private val coroutineScope: CoroutineScope) : MediaPlayer {
     private val log: Logger = Logger("de.connect2x.trixnity.messenger.media.WebMediaPlayer")
-    internal val currentItemPlaying: MutableStateFlow<AbstractMediaItem?> = MutableStateFlow(null)
-    internal val playerMutex: Mutex = Mutex()
-
-    private val audioContext: AudioContext = AudioContext()
-
-    override val playingItem: StateFlow<MediaPlayer.Item?> = currentItemPlaying.asStateFlow()
-
-    init {
+    private val removeAudioContextErrorHandler: () -> Unit =
         audioContext.addEventHandler(
             type = Event.ERROR,
             handler = EventHandler { log.error { "Unexpected media player error" } },
         )
+    internal val currentItemPlaying: MutableStateFlow<AbstractMediaItem?> = MutableStateFlow(null)
+    internal val playerMutex: Mutex = Mutex()
 
+    override val playingItem: StateFlow<MediaPlayer.Item?> = currentItemPlaying.asStateFlow()
+
+    init {
         handleMediaSessionActions()
     }
 
@@ -120,7 +122,8 @@ class WebMediaPlayer(private val coroutineScope: CoroutineScope) : MediaPlayer {
     }
 
     override fun close() {
-        audioContext.closeAsync()
+        removeMediaSessionActionHandlers()
+        removeAudioContextErrorHandler()
     }
 
     private suspend fun initAudio(tempFile: PlatformMedia.TemporaryFile): Audio {
@@ -203,5 +206,23 @@ class WebMediaPlayer(private val coroutineScope: CoroutineScope) : MediaPlayer {
                 }
             }
             .launchIn(coroutineScope)
+    }
+
+    private fun removeMediaSessionActionHandlers() {
+        val allMediaSessionActions =
+            listOf(
+                MediaSessionAction.nexttrack,
+                MediaSessionAction.pause,
+                MediaSessionAction.play,
+                MediaSessionAction.previoustrack,
+                MediaSessionAction.seekbackward,
+                MediaSessionAction.seekforward,
+                MediaSessionAction.seekto,
+                MediaSessionAction.skipad,
+                MediaSessionAction.stop,
+            )
+        allMediaSessionActions.forEach { mediaSessionAction ->
+            navigator.mediaSession.setActionHandler(action = mediaSessionAction, handler = null)
+        }
     }
 }