+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDK: search providers are registered with a factory, so that they can be disabled for certain accounts
 - SDK: Added check for downloadsDisabled before downloading file
 - SDK: granular download limits for syncing media files in `MatrixMessengerConfiguration`
+- SDK,UI,WEB: Experimental voice messages
 
 ### Changed
 
@@ -32,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDK: (**breaking change**) Changed `GetEventReactions.EventReaction` to use `EventIdOrTransactionId` instead of `EventId` in order to support reactions in outbox
 - DEPENDENCY: update compose to 1.11.1
 - SDK: files aren't automatically downloaded when they exceed the configurable limit
+- SDK: Changed naming and removed functions of the `AudioRecorder` API including its view model  
+- SDK: The `AudioRecorder` API now only returns a reference to an item inside the media store instead of returning a `PlatformMedia` directly 
 
 ### Deprecated